pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "gfal"
dynamic = ["version"]
description = "Grid File Access Library — Python rewrite of gfal2, based on fsspec (HTTP/HTTPS and XRootD)"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
    { name = "Luis Antonio Obis Aparicio", email = "luis.obis@cern.ch" },
]
dependencies = [
    "fsspec>=2023.1.0",
    "fsspec-xrootd>=0.5.1",
    "aiohttp>=3.8.0",
    "requests>=2.28.0",
    "truststore; python_version >= '3.10'",
    "textual>=0.50.0",
    "rich>=13.0.0",
    "rich-click>=1.7.0",
]

[project.scripts]
gfal = "gfal.cli.shell:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "pytest-xdist>=3.0",
    "pre-commit>=3.0",
    "codespell>=2.2.0",
]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/gfal"]

[tool.hatch.build.hooks.vcs]
version-file = "src/gfal/_version.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-n 4"
markers = [
    "integration: tests that require network access to external services",
    "xrootd: tests that require a local XRootD server (started automatically by the fixture)",
]

[tool.ruff]
src = ["src", "tests"]
target-version = "py39"
exclude = ["src/gfal/_version.py"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "C4",  # flake8-comprehensions
    "PTH", # use pathlib over os.path (advisory)
]
ignore = [
    "E501",   # line too long – handled by the formatter
    "UP045",  # use `X | None` for type annotations (wait: we need 3.9 support!)
]

[tool.ruff.lint.isort]
known-first-party = ["gfal"]

[tool.codespell]
skip = ".git,.venv,*.egg-info"
ignore-words-list = "cern,xrootd,fsspec,gfal,adler,aiohttp,thirdparty,scitag"
